+++ src/repositories/user.repository.ts
@@ -1,5 +1,5 @@
 import { AbstractRepository } from './abstract.repository';
-import { UserDocument } from '../modules/user/user.interface';
+import { UserDocument } from '../modules/user/user.document';
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';